+++ src/MultiSigWallet.sol
@@ -1,17 +1,19 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-contract MultiSigWallet{
-    event Deposit(address indexed sender, uint256 amount, uint balance);
-    event SubmitTransaction(address indexed owner, uint256 indexed txIndex, address indexed to, uint value, bytes data);
+contract MultiSigWallet {
+    event Deposit(address indexed sender, uint256 amount, uint256 balance);
+    event SubmitTransaction(
+        address indexed owner, uint256 indexed txIndex, address indexed to, uint256 value, bytes data
+    );
     event ConfirmTransaction(address indexed owner, uint256 indexed txIndex);
     event RevokeConfirmation(address indexed owner, uint256 indexed txIndex);
     event ExecuteTransaction(address indexed owner, uint256 indexed txIndex);
 
     address[] public owners;
     mapping(address => bool) public isOwner;
-    uint public numConfirmationsRequired;
-    mapping (uint256 => mapping(address => bool)) public isConfirmed;
+    uint256 public numConfirmationsRequired;
+    mapping(uint256 => mapping(address => bool)) public isConfirmed;
     Transaction[] public transactions;
 
     struct Transaction {
@@ -27,25 +29,28 @@ contract MultiSigWallet{
         _;
     }
 
-    modifier txExists (uint256 _txIndex) {
-        require (_txIndex < transactions.length, "Tx does not exists");
+    modifier txExists(uint256 _txIndex) {
+        require(_txIndex < transactions.length, "Tx does not exists");
         _;
     }
 
-    modifier notExecuted (uint256 _txIndex) {
-        require (!transactions[_txIndex].executed, "Tx already executed");
-        _;        
+    modifier notExecuted(uint256 _txIndex) {
+        require(!transactions[_txIndex].executed, "Tx already executed");
+        _;
     }
 
-    modifier notConfirmed (uint256 _txIndex) {
+    modifier notConfirmed(uint256 _txIndex) {
         require(!isConfirmed[_txIndex][msg.sender], "Tx already confirmed");
         _;
     }
 
     constructor(address[] memory _owners, uint256 _numConfirmationsRequired) {
         require(_owners.length > 0, "Owners required");
-        require(_numConfirmationsRequired > 0 && _numConfirmationsRequired <= _owners.length, "Invalid number of required confirmations");
-        
+        require(
+            _numConfirmationsRequired > 0 && _numConfirmationsRequired <= _owners.length,
+            "Invalid number of required confirmations"
+        );
+
         for (uint256 i = 0; i < _owners.length; i++) {
             address owner = _owners[i];
             require(owner != address(0), "Invalid owner");
@@ -63,34 +68,34 @@ contract MultiSigWallet{
 
     function submitTransaction(address _to, uint256 _value, bytes memory _data) public onlyOwner {
         uint256 txIndex = transactions.length;
-        transactions.push( Transaction ({
-            to: _to,
-            value: _value,
-            data: _data,
-            executed: false,
-            numConfirmations: 0
-        }));
+        transactions.push(Transaction({to: _to, value: _value, data: _data, executed: false, numConfirmations: 0}));
         emit SubmitTransaction(msg.sender, txIndex, _to, _value, _data);
     }
 
-    function confirmTransaction(uint256 _txIndex) public onlyOwner notExecuted(_txIndex) notConfirmed(_txIndex){
+    function confirmTransaction(uint256 _txIndex) public onlyOwner notExecuted(_txIndex) notConfirmed(_txIndex) {
         Transaction storage transaction = transactions[_txIndex];
         transaction.numConfirmations += 1;
         isConfirmed[_txIndex][msg.sender] = true;
         emit ConfirmTransaction(msg.sender, _txIndex);
     }
 
-    function executeTransaction(uint256 _txIndex) public onlyOwner txExists(_txIndex) notExecuted(_txIndex) returns (bool) {
+    function executeTransaction(uint256 _txIndex)
+        public
+        onlyOwner
+        txExists(_txIndex)
+        notExecuted(_txIndex)
+        returns (bool)
+    {
         Transaction storage transaction = transactions[_txIndex];
-        require (transaction.numConfirmations >= numConfirmationsRequired, "Can not execute tx");
+        require(transaction.numConfirmations >= numConfirmationsRequired, "Can not execute tx");
         transaction.executed = true;
-        (bool success, ) = transaction.to.call{value: transaction.value}(transaction.data);
-        require (success, "Tx failed");
+        (bool success,) = transaction.to.call{value: transaction.value}(transaction.data);
+        require(success, "Tx failed");
         emit ExecuteTransaction(msg.sender, _txIndex);
         return success;
     }
 
-    function revokeConfirmation(uint _txIndex) public onlyOwner txExists(_txIndex) notExecuted(_txIndex) {
+    function revokeConfirmation(uint256 _txIndex) public onlyOwner txExists(_txIndex) notExecuted(_txIndex) {
         Transaction storage transaction = transactions[_txIndex];
         require(isConfirmed[_txIndex][msg.sender], "Tx not confirmed");
         transaction.numConfirmations -= 1;
@@ -102,15 +107,19 @@ contract MultiSigWallet{
         return owners;
     }
 
-    function getTransactionCount() public view returns (uint) {
+    function getTransactionCount() public view returns (uint256) {
         return transactions.length;
     }
 
-    function getBalance() public view returns (uint) {
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
 
-    function getTransaction(uint256 _txIndex) public view returns (address to, uint256 value, bytes memory data, bool executed, uint256 numConfirmations){
+    function getTransaction(uint256 _txIndex)
+        public
+        view
+        returns (address to, uint256 value, bytes memory data, bool executed, uint256 numConfirmations)
+    {
         Transaction storage transaction = transactions[_txIndex];
         return (transaction.to, transaction.value, transaction.data, transaction.executed, transaction.numConfirmations);
     }
